# docs/api_surface/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Campus Resource Hub API
  version: 1.0.0
  description: |
    RESTful API for the Campus Resource Hub - A university resource booking and management system.
    
    **Implementation Status:**
    - ✅ existing: Backend implemented and functional
    - ⚠️ partial: Partially implemented
    - ❌ missing: Not yet implemented
    
    **Authentication:** Session-based cookies (Flask-Login)
    
    **Base URL:** `/api`
  contact:
    name: Campus Resource Hub Team
    email: support@campusresourcehub.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://production-domain.com/api
    description: Production server (TBD)
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie for authentication
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection token required for all state-changing operations (POST, PUT, PATCH, DELETE).
        Obtain token from GET /api/auth/csrf-token endpoint.
        Token lifetime: 1 hour
        Alternative header name: X-CSRFToken
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [student, staff, admin]
        department:
          type: string
        status:
          type: string
          enum: [active, pending, suspended]
        created_at:
          type: string
          format: date-time
      required: [id, name, email, role]
    Resource:
      type: object
      properties:
        id:
          type: integer
        owner_id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [study_room, lab, equipment, facility, vehicle, technology, sports, event_space, other]
        location:
          type: string
        capacity:
          type: integer
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
        requires_approval:
          type: boolean
        average_rating:
          type: number
          format: float
        review_count:
          type: integer
        created_at:
          type: string
          format: date-time
      required: [id, title, status]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
  responses:
    UnauthorizedError:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests (will be 0)
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too Many Requests"
              message:
                type: string
                example: "Rate limit exceeded. Please try again later."
              retry_after:
                type: string
                example: "60 seconds"
              status:
                type: integer
                example: 429
paths:
...
  /resources:
    get:
      tags: [Resources]
      summary: List resources with pagination and filters
      x-status: existing
      description: |
        Get a paginated list of resources with optional filters.
        Rate limit: 50 per hour (default)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of resources
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per hour
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when limit resets
          content:
            application/json:
              schema:
                type: object
                properties:
                  resources:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resource'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Resources]
      summary: Create new resource
      x-status: existing
      description: |
        Create a new resource listing.
        Rate limit: 20 per hour
      security:
        - SessionCookie: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                location:
                  type: string
                capacity:
                  type: integer
                requires_approval:
                  type: boolean
                  default: true
                status:
                  type: string
                  enum: [draft, published]
                  default: draft
      responses:
        '201':
          description: Resource created
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  resource:
                    $ref: '#/components/schemas/Resource'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: CSRF validation failed or forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /admin/analytics:
    get:
      tags: [Admin]
      summary: Get system analytics (admin only)
      x-status: existing
      security:
        - SessionCookie: []
      responses:
        '200':
          description: System analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: object
                  resources:
                    type: object
                  bookings:
                    type: object
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin only)
      x-status: existing
      security:
        - SessionCookie: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /admin/users/{id}/role:
    put:
      tags: [Admin]
      summary: Update user role (admin only)
      x-status: existing
      security:
        - SessionCookie: []
        - CSRFToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [student, staff, admin]
      responses:
        '200':
          description: Role updated
  /admin/reviews/flagged:
    get:
      tags: [Admin]
      summary: Get flagged reviews (admin only)
      x-status: existing
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Flagged reviews
...